+++ src/table.ts
@@ -1376,13 +1376,6 @@ class Table extends common.ServiceObject {
       delete metadata.jobPrefix;
     }
 
-    if (
-      metadata.hasOwnProperty('sourceFormat') &&
-      fileTypes.indexOf(metadata.sourceFormat!) < 0
-    ) {
-      throw new Error(`Source format not recognized: ${metadata.sourceFormat}`);
-    }
-
     const dup = streamEvents(duplexify());
 
     dup.once('writing', () => {
@@ -1402,7 +1395,9 @@ class Table extends common.ServiceObject {
           } as {},
           request: {
             uri: format('{base}/{projectId}/jobs', {
-              base: 'https://www.googleapis.com/upload/bigquery/v2/projects',
+              base: `https://${
+                this.bigQuery.apiEndpoint
+              }/upload/bigquery/v2/projects`,
               projectId: this.bigQuery.projectId,
             }),
           },